+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log (vscode-powertools)
 
 
+## 0.12.0 (January 23rd, 2019; global buttons and jobs)
+
+* [buttons](https://github.com/egodigital/vscode-powertools/wiki/Buttons) can be defined [globally](https://github.com/egodigital/vscode-powertools/wiki/Global%20Settings) now (`CTRL + ,`)
+
 ## 0.11.0 (January 23rd, 2019; npm update & bug fixes)
 
 * bug fix: there was no auto creation of `.vscode-powertools` sub folder in home directory, when using [store](https://egodigital.github.io/vscode-powertools/api/classes/_stores_.userstore.html) in scripts, e.g. ... s. [issue #7](https://github.com/egodigital/vscode-powertools/issues/7)